# Banking stage: Deserialize packets only once