@taozhu-chicago 2022-04-13
> @taozhu-chicago I know you have a lot more changes pending for this part of the code -- but this maybe worthwhile to get into v1.10 since the bigger changes are likely to only come later?

Yea, the changes in pipeline will eventually remove/replace `filter_unprocessed_packets_at_end_of_slot` and `transactions_from_packets`, but looks like it's taking time. Glad to get 6% boost for the time being, thank you!